requirements.txt
fastapi
uvicorn
torch
torchvision
timm
pillow
opencv-python-headless
numpy
python-multipart